+++ cmd/loom/config.go
@@ -168,7 +168,7 @@ func defaultGenesis(validator *loom.Validator) (*genesis, error) {
 		Params: &karma.Params{
 			MaxKarma: 10000,
 			OraclePublicAddress: "irn38gFRpNOzoySXECh5JZVoPm1Hw6UAqCdeqv4IQlM=", // change to real oracle key
-			Sources: map[string]float64{
+			Sources: map[string]int64{
 				"sms": 1.0, //default sources and values
 				"oauth": 3.0, //default sources and values
 				"token": 5.0, //default sources and values
